Unref(): always decrement ref count atomically
Unref() included an optimization for the case where the refcount is
already 1. In that case it avoided the atomic decrement to 0 and just
proceeded with the destruction of the object.

But that optimization causes a problem on POWER. With debug enabled the
destruction of dense_graph in SimpleGraphExecutionState::BuildGraph()
(either during std::move() or immediately after when dense_graph goes
out of scope) leads down a path that ultimately assert()s in
~RefCounted():

F ./tensorflow/core/lib/core/refcount.h:79] Check failed: ref_.load() == 0 (1 vs. 0)

(gdb) where
[...]
3  tensorflow::core::RefCounted::~RefCounted (this=0x15e729f0, ...)
   at ./tensorflow/core/lib/core/refcount.h:79
4  tensorflow::Node::Properties::~Properties (this=0x15e729f0, ...)
   at tensorflow/core/graph/graph.cc:166
5  tensorflow::Node::Properties::~Properties (this=0x15e729f0, ...)
   at tensorflow/core/graph/graph.cc:166
6  tensorflow::core::RefCounted::Unref (this=0x15e729f0)
   at ./tensorflow/core/lib/core/refcount.h:93
7  tensorflow::Node::~Node (this=0x15e2aca0, ...)
   at tensorflow/core/graph/graph.cc:59
8  tensorflow::Graph::~Graph (this=0x15e72790, ...)
   at tensorflow/core/graph/graph.cc:197
9  tensorflow::SimpleClientGraph::~SimpleClientGraph (this=0x15e72790, ...)
   at ./tensorflow/core/common_runtime/simple_graph_execution_state.h:49
10 std::default_delete<tensorflow::SimpleClientGraph>::operator() (...)
   at /usr/include/c++/5/bits/unique_ptr.h:76
11 std::unique_ptr<tensorflow::SimpleClientGraph, std::default_delete<tensorflow::SimpleClientGraph> >::reset (...)
   at /usr/include/c++/5/bits/unique_ptr.h:344
12 std::unique_ptr<tensorflow::SimpleClientGraph, std::default_delete<tensorflow::SimpleClientGraph> >::operator=(...) (...)
   at /usr/include/c++/5/bits/unique_ptr.h:251
13 tensorflow::SimpleGraphExecutionState::BuildGraph (...)
   at tensorflow/core/common_runtime/simple_graph_execution_state.cc:226
14 tensorflow::DirectSession::CreateGraphs (...)
   at tensorflow/core/common_runtime/direct_session.cc:866

When built without debug the assert does not trigger, but the program
suffers seemingly-random memory corruption usually ending in a crash of
some sort.

Changing Unref() to always atomicly decrement the ref count fixes the
assert. It seems to prevent the random crashes for low thread counts
(NumSchedulableCPUs() <= 8 no longer crashes, >= 16 still crashes).