Unref(): always decrement ref count atomically
Unref() included an optimization for the case where the refcount is
already 1. In that case it avoided the atomic decrement to 0 and just
proceeded with the destruction of the object.

But that causes a crash on POWER. With debug enabled the destruction of
dense_graph in SimpleGraphExecutionState::BuildGraph() (either during
std::move() or immediately after when dense_graph goes out of scope)
leads down a path that ultimately assert()s in ~RefCounted():

F ./tensorflow/core/lib/core/refcount.h:79] Check failed: ref_.load() == 0 (1 vs. 0)

(gdb) where
[...]
        at ./tensorflow/core/lib/core/refcount.h:79
        at tensorflow/core/graph/graph.cc:166
        at tensorflow/core/graph/graph.cc:166
        at ./tensorflow/core/lib/core/refcount.h:93
        at tensorflow/core/graph/graph.cc:59
        at tensorflow/core/graph/graph.cc:197
        at ./tensorflow/core/common_runtime/simple_graph_execution_state.h:49
        at /usr/include/c++/5/bits/unique_ptr.h:76
        at /usr/include/c++/5/bits/unique_ptr.h:344
        at /usr/include/c++/5/bits/unique_ptr.h:251
        at tensorflow/core/common_runtime/simple_graph_execution_state.cc:226

When built without debug the assert does not trigger, but the program
suffers seemingly-random memory corruption usually ending in a crash of
some sort.

Changing Unref() to always do an atomic decrement of the ref count
prevents both the assert and the crash.